move _convert_to_content
