+++ src/mcp/server/fastmcp/server.py
@@ -9,7 +9,6 @@ from contextlib import (
     AbstractAsyncContextManager,
     asynccontextmanager,
 )
-from itertools import chain
 from typing import Any, Generic, Literal
 
 import anyio
@@ -38,7 +37,6 @@ from mcp.server.fastmcp.prompts import Prompt, PromptManager
 from mcp.server.fastmcp.resources import FunctionResource, Resource, ResourceManager
 from mcp.server.fastmcp.tools import Tool, ToolManager
 from mcp.server.fastmcp.utilities.logging import configure_logging, get_logger
-from mcp.server.fastmcp.utilities.types import Image
 from mcp.server.lowlevel.helper_types import ReadResourceContents
 from mcp.server.lowlevel.server import LifespanResultT
 from mcp.server.lowlevel.server import Server as MCPServer
@@ -54,7 +52,6 @@ from mcp.types import (
     AnyFunction,
     ContentBlock,
     GetPromptResult,
-    TextContent,
     ToolAnnotations,
 )
 from mcp.types import Prompt as MCPPrompt
@@ -966,28 +963,6 @@ class FastMCP:
         except Exception as e:
             logger.error(f"Error getting prompt {name}: {e}")
             raise ValueError(str(e))
-
-
-def _convert_to_content(
-    result: Any,
-) -> Sequence[ContentBlock]:
-    """Convert a result to a sequence of content objects."""
-    if result is None:
-        return []
-
-    if isinstance(result, ContentBlock):
-        return [result]
-
-    if isinstance(result, Image):
-        return [result.to_image_content()]
-
-    if isinstance(result, list | tuple):
-        return list(chain.from_iterable(_convert_to_content(item) for item in result))  # type: ignore[reportUnknownVariableType]
-
-    if not isinstance(result, str):
-        result = pydantic_core.to_json(result, fallback=str, indent=2).decode()
-
-    return [TextContent(type="text", text=result)]
 
 
 class Context(BaseModel, Generic[ServerSessionT, LifespanContextT, RequestT]):
